modifying taskscontroller

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(3);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('tasks.index',[
-            'tasks' => $tasks,
-            ]);
+            return view('tasks.index', $data);
     }
 
 
@@ -30,7 +37,9 @@ class TasksController extends Controller
      */
     public function create()
     {
+        if (\Auth::check()) {
         $task = new Task;
+        }
         
         return view('tasks.create', [
             'task' => $task,
@@ -45,7 +54,7 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {   
-        $auths = Auth::user();
+        if (\Auth::check()) {
         $this->validate($request, [
             'content' => 'required|max:191',
             'status' => 'required|max:10',
@@ -56,6 +65,7 @@ class TasksController extends Controller
         $task->content = $request->content;
         $task->status = $request->status; 
         $task->save();
+        }
         
         return redirect('/');
     }
@@ -81,7 +91,9 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
+        if (\Auth::check()) {
         $task = Task::find($id);
+        }
         
         return view('tasks.edit', ['task'=>$task
         ]);
@@ -101,11 +113,13 @@ class TasksController extends Controller
         'status' => 'required|max:10',
         ]);
         
-       $task = Task::find($id);
-       $task->content = $request->content;
-       $task->status = $request->status; 
-       $task->save();
-       
+        if (\Auth::check()) {
+        $task = Task::find($id);
+        $task->content = $request->content;
+        $task->status = $request->status; 
+        $task->save();
+        }
+        
        return redirect('/');
     }
 
@@ -117,9 +131,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-       $task = Task::find($id);
-       $task->delete();
-       
-       return redirect ('/');
+        $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+        $task->delete();
+        }
+        
+       return redirect('/');
     }
 }
